Rename resource model in OperatorManagement `AttendableResourceModel`
