+++ src/Moryx.Operators.Web/src/app/api/fn/operator-management/get-resources.ts
@@ -6,7 +6,7 @@ import { filter, map } from 'rxjs/operators';
 import { StrictHttpResponse } from '../../strict-http-response';
 import { RequestBuilder } from '../../request-builder';
 
-import { OperatorsResourceModel as MoryxOperatorsEndpointsResourceModel } from '../../models/Moryx/Operators/Endpoints/operators-resource-model';
+import { AttendableResourceModel as MoryxOperatorsEndpointsResourceModel } from '../../models/Moryx/Operators/Endpoints/attendable-resource-model';
 
 export interface GetResources$Params {
   operatorIdentifier: string;
