+++ src/Moryx.Operators.Web/src/app/api/models/Moryx/Operators/Endpoints/attendable-resource-model.ts
@@ -1,6 +1,6 @@
 /* tslint:disable */
 /* eslint-disable */
-export interface OperatorsResourceModel {
+export interface AttendableResourceModel {
   id?: number | null;
   name?: string | null;
   requiredSkills?: Array<string> | null;
