+++ src/Moryx.Operators.Web/src/app/models/operator-view-model.ts
