+++ src/Moryx.Operators.Web/src/app/models/workstation-view-model.ts
@@ -1,10 +1,10 @@
-import { OperatorsResourceModel } from "../api/models/Moryx/Operators/Endpoints/operators-resource-model";
+import { AttendableResourceModel } from "../api/models/Moryx/Operators/Endpoints/attendable-resource-model";
 import { OperatorViewModel } from "./operator-view-model";
 
 export class WorkstationViewModel  {
 
     constructor(
-        private _data : OperatorsResourceModel ,
+        private _data : AttendableResourceModel ,
     )
         {}
